PWN-3247 - make TransactionImage.kt to use as common item to show  * attrs.xml to use for custom sizes in progress (need to adjust for bottomSheet)
