+++ app/src/main/res/layout/item_transaction.xml
@@ -1,39 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@color/backgroundPrimary"
     android:orientation="horizontal">
 
-    <FrameLayout
-        android:layout_width="46dp"
-        android:layout_height="46dp"
+    <org.p2p.wallet.common.ui.widget.TransactionImage
+        android:id="@+id/transactionTokenImageView"
+        android:layout_width="@dimen/history_transaction_image_size"
+        android:layout_height="@dimen/history_transaction_image_size"
         android:layout_marginVertical="12dp"
-        android:layout_marginStart="16dp">
-
-        <org.p2p.wallet.common.ui.widget.RoundedImageView
-            android:id="@+id/transactionTokenImageView"
-            android:layout_width="@dimen/history_transaction_image_size"
-            android:layout_height="@dimen/history_transaction_image_size"
-            android:background="@drawable/bg_token_rounded"
-            android:padding="12dp"
-            android:src="@drawable/ic_transaction_send"
-            app:cornerRadius="12dp"
-            tools:visibility="visible" />
-
-        <ImageView
-            android:id="@+id/transactionStatus"
-            android:layout_width="18dp"
-            android:layout_height="18dp"
-            android:layout_gravity="bottom|end"
-            android:visibility="gone"
-            tools:src="@drawable/ic_state_pending"
-            tools:visibility="visible" />
-
-    </FrameLayout>
+        android:layout_marginStart="16dp" />
 
     <include
         android:id="@+id/transactionData"
