+++ app/src/main/res/layout/widget_transaction_image.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:parentTag="android.widget.FrameLayout">
+
+    <org.p2p.wallet.common.ui.widget.RoundedImageView
+        android:id="@+id/transactionTokenImageView"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@drawable/bg_token_rounded"
+        android:padding="12dp"
+        android:layout_margin="2dp"
+        android:src="@drawable/ic_transaction_send"
+        app:cornerRadius="12dp"
+        tools:visibility="visible" />
+
+    <ImageView
+        android:id="@+id/transactionStatus"
+        android:layout_width="18dp"
+        android:layout_height="18dp"
+        android:layout_gravity="bottom|end"
+        android:visibility="gone"
+        tools:src="@drawable/ic_state_pending"
+        tools:visibility="visible" />
+
+</merge>
